requirements.txt
playwright
pyyaml
httpx
sqlalchemy
asyncpg
fastapi
uvicorn[standard]
jinja2
rapidfuzz
